rename cmds folder to command_storage
