+++ command_storage/views/_read_cli.py
@@ -5,10 +5,10 @@ from pathlib import Path
 import typer
 from tabulate import tabulate
 
-from cmds.controller.app import get_cmds
-from cmds.initializer import app
-from cmds.models.enums import arguments as arguments_enums
-from cmds.models.enums import error as error_enums
+from command_storage.controller.app import get_cmds
+from command_storage.initializer import app
+from command_storage.models.enums import arguments as arguments_enums
+from command_storage.models.enums import error as error_enums
 
 DEFAULT_FILE_LOCATION = Path().joinpath(f"command_storage_export_{datetime.now()}.json")
